+++ src/components/Header/Header.tsx
@@ -1,4 +1,12 @@
-import { createStyles, Group, Image, Stack, Tabs, Text } from "@mantine/core";
+import {
+  Badge,
+  createStyles,
+  Group,
+  Image,
+  Stack,
+  Tabs,
+  Text,
+} from "@mantine/core";
 import { useLocation, useNavigate } from "react-router-dom";
 
 import { ROUTES } from "../../constants";
@@ -12,6 +20,14 @@ const tabs = [
   ROUTES.GALAXY_BUILDER,
   ROUTES.OTHER_TOOLS,
 ];
+
+const isProd = RUNTIME_ENV === "prod";
+
+const BADGE_COLOR_MAP = {
+  local: "pink",
+  dev: "orange",
+  stage: "cyan",
+};
 
 const useStyles = createStyles((theme) => ({
   header: {
@@ -53,6 +69,11 @@ export default function Header() {
             <Text size="xl" weight={700}>
               RMG Utils for Stellaris
             </Text>
+            {!isProd && (
+              <Badge color={BADGE_COLOR_MAP[RUNTIME_ENV]} variant="outline">
+                {RUNTIME_ENV}
+              </Badge>
+            )}
           </Group>
           <ColorSchemeTogle />
         </Group>