+++ src/vite-env.d.ts
@@ -1,5 +1,6 @@
 /// <reference types="vite/client" />
 declare const APP_VERSION: string;
+declare const RUNTIME_ENV: "local" | "dev" | "stage" | "prod";
 
 interface ImportMetaEnv {
   readonly VITE_NETLIFY_FUNCTIONS_LOCAL_SERVER: string;