+++ vite.config.ts
@@ -2,6 +2,17 @@ import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import { VitePWA } from "vite-plugin-pwa";
 
+const env = process.env.RUNTIME_ENV ?? "local";
+
+const PWA_TITLE_MAP = {
+  local: "RMG Utils (local)",
+  prod: "RMG Utilities for Stellaris",
+  stage: "RMG Utils (stage)",
+  dev: "RMG Utils (dev)",
+};
+
+const PWA_ICONS_TYPE = env === "prod" ? "" : "_" + env;
+
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
@@ -9,19 +20,19 @@ export default defineConfig({
     VitePWA({
       registerType: "autoUpdate",
       manifest: {
-        name: "RMG Utilities for Stellaris",
+        name: PWA_TITLE_MAP[env],
         short_name: "RMG Utils",
         description:
           "RMG Utils is an application that RMG is using to expediate some parts of the mod-making process.",
         theme_color: "#25262b",
         icons: [
           {
-            src: "/assets/images/rmg_logo_192.png",
+            src: `/assets/images/rmg_logo_192${PWA_ICONS_TYPE}.png`,
             sizes: "192x192",
             type: "image/png",
           },
           {
-            src: "/assets/images/rmg_logo_512.png",
+            src: `/assets/images/rmg_logo_512${PWA_ICONS_TYPE}.png`,
             sizes: "512x512",
             type: "image/png",
           },
@@ -46,5 +57,6 @@ export default defineConfig({
   ],
   define: {
     APP_VERSION: JSON.stringify(process.env.npm_package_version),
+    RUNTIME_ENV: JSON.stringify(env),
   },
 });